add basic enemy controller script
