+++ Assets/Scripts/Enemies/EnemyController.cs
@@ -1,0 +1,36 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class EnemyController : MonoBehaviour
+{
+    public GameObject player;
+    public PlayerData playerData;
+    public float speed;
+    public float damage;
+    public float baseHealth;
+    public float health;    
+    void Start()
+    {
+        health = baseHealth;
+    }
+
+    void Update()
+    {
+        if (health <= 0)
+        {
+            OnDeath();
+        }
+    }
+
+    public void TakeDamage(float damageAmount)
+    {
+        health -= damageAmount;
+    }
+
+    void OnDeath()
+    {
+        Destroy(gameObject);
+        //tba maybe item drops
+    }   
+}
