+++ Assets/Scripts/Enemies/EnemyController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e55c3ebab924ff341ab5bce5b6551a63
+guid: c5e1ccc94e931db4a8a356343bedec86
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
