+++ Assets/Scripts/Weapons/Flamethrower.cs
@@ -62,7 +62,7 @@ only the UseWeapon method different
                 //deal damage
                 try
                 {
-                    hit.collider.gameObject.GetComponent<EnemyController>().health -= damage;
+                    hit.collider.gameObject.GetComponent<EnemyController>().TakeDamage(damage);
                 }
                 catch (Exception e)
                 {
