+++ Assets/Scripts/Weapons/TaserGun.cs
@@ -44,7 +44,7 @@ public class TaserGun : Weapon
                 // deal damage
                 try
                 {
-                    hit.collider.gameObject.GetComponent<EnemyController>().health -= damage;
+                    hit.collider.gameObject.GetComponent<EnemyController>().TakeDamage(damage);
                 }
                 catch (Exception e)
                 {
